Add Preparation phase (#1866)
* Moved all config changing plugin into new prepearation phase

* Change remember rejected into what it was before config hash refactor

* Changed preparation to prepare

* Skipping prepare phase on rerun

+++ flexget/plugins/filter/series.py
@@ -2058,6 +2058,10 @@ class SeriesDBManager(FilterSeriesBase):
 
     @plugin.priority(0)
     def on_task_start(self, task, config):
+        # Only operate if task changed
+        if not task.config_modified:
+            return
+
         # Clear all series from this task
         with Session() as session:
             session.query(SeriesTask).filter(SeriesTask.name == task.name).delete()
